# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tokenisation Platform API
  description: |
    API for managing security token issuance, compliance, and distribution.

    ## Authentication
    All endpoints require authentication via Bearer token in the Authorization header.

    ## Rate Limits
    - Standard tier: 100 requests per minute
    - Premium tier: 1000 requests per minute

    ## Webhooks
    Configure webhooks in the dashboard to receive real-time notifications for:
    - Token transfers
    - KYC status changes
    - Dividend distributions
    - Compliance events
  version: 1.0.0
  contact:
    name: API Support
    email: support@tokenisation.example.com
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.tokenisation.example.com/v1
    description: Production
  - url: https://api-staging.tokenisation.example.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
...
paths:
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expiresIn:
                    type: integer

  # ==================== TOKENS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []